Move each rating system into its own file
