+++ src/lib/ratings/openskillRatingSystem.ts
@@ -1,0 +1,90 @@
+import { ordinal, rate, rating } from "openskill";
+import { type Options, type Rating } from "openskill/dist/types";
+import { isDefined } from "../utils";
+import {
+  type MatchWithRatings,
+  type PlayerWithRating,
+  type RatingSystem,
+} from "./rating";
+
+export type OpenskillRating = Rating;
+
+export function openskill(options?: Options): RatingSystem<OpenskillRating> {
+  const selectedOptions: Options = options ?? {
+    mu: 1000, // skill level, higher is better
+    sigma: 500, // certainty, lower is more certain
+    tau: 0.3, // tau prevents model from getting too certain about a players skill level
+    z: 2, // used in calculation of ordinal `my - z * sigma`
+  };
+
+  return {
+    defaultRating: rating(selectedOptions),
+
+    rateMatch(
+      match: MatchWithRatings<OpenskillRating>,
+    ): PlayerWithRating<OpenskillRating>[] {
+      const whiteTeam = [
+        match.whitePlayerOne.rating,
+        match.whitePlayerTwo?.rating,
+      ].filter(isDefined);
+
+      const blackTeam = [
+        match.blackPlayerOne.rating,
+        match.blackPlayerTwo?.rating,
+      ].filter(isDefined);
+
+      // Lower is better
+      // It makes a difference if the ranking is zero or non-zero, not sure why 🤷
+      const outcomeRanking = {
+        White: [1, 2],
+        Black: [2, 1],
+        Draw: [1, 1],
+      }[match.result];
+
+      const [
+        [whitePlayerOneNewRating, whitePlayerTwoNewRating],
+        [blackPlayerOneNewRating, blackPlayerTwoNewRating],
+      ] = rate([whiteTeam, blackTeam], {
+        ...selectedOptions,
+        rank: outcomeRanking,
+      });
+
+      const result: PlayerWithRating<OpenskillRating>[] = [
+        {
+          player: match.whitePlayerOne.player,
+          rating: whitePlayerOneNewRating,
+        },
+        {
+          player: match.blackPlayerOne.player,
+          rating: blackPlayerOneNewRating,
+        },
+      ].filter((x) => isDefined(x.player));
+
+      if (match.whitePlayerTwo) {
+        result.push({
+          player: match.whitePlayerTwo.player,
+          rating: whitePlayerTwoNewRating,
+        });
+      }
+
+      if (match.blackPlayerTwo) {
+        result.push({
+          player: match.blackPlayerTwo?.player,
+          rating: blackPlayerTwoNewRating,
+        });
+      }
+
+      return result;
+    },
+
+    toNumber(rating: OpenskillRating) {
+      return Math.floor(ordinal(rating, selectedOptions));
+    },
+
+    equals(a: OpenskillRating | undefined, b: OpenskillRating | undefined) {
+      if (a === undefined && b === undefined) return true;
+      if (a === undefined || b === undefined) return false;
+      return a.sigma === b.sigma && a.mu === b.mu;
+    },
+  };
+}
