--- src/types/elo.d.ts
@@ -1,3 +1,0 @@
-export interface EloConfig {
-  eloFloor: number;
-}
